requirements.txt
python-pptx==0.6.6
pandas>=0.18.0
openpyxl>=2.4.7
six